Factor out MemoryAPI header
